requirements.txt
neat-python
pygame-2.4.0